+++ tests/formatter/input-output-pairs/simple-atom-expressions.out.gd
@@ -18,3 +18,5 @@ func foo():
 	var n = x
 	var d = ^"../Some/Stuff"
 	var e = &"xxx"
+	var xa = $%UniqueNodeName
+	var xb = $%UniqueNodeName/Xyz